feat: Implement streaming for the dashboard page
