+++ app/dashboard/(overview)/loading.tsx
@@ -1,0 +1,7 @@
+import DashboardSkeleton from '@/app/ui/skeletons';
+
+function Loading() {
+  return <DashboardSkeleton />;
+}
+
+export default Loading;
